+++ src/test/sigopcount_tests.cpp
@@ -2,7 +2,7 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include "validation.h"
+#include "consensus/tx_verify.h"
 #include "pubkey.h"
 #include "key.h"
 #include "script/script.h"
